+++ minari/__init__.py
@@ -1,5 +1,5 @@
-from minari.minari_dataset import (
-    MinariDataset,
+from minari.minari_dataset import MinariDataset
+from minari.utils import (
     combine_datasets,
     create_dataset_from_buffers,
     create_dataset_from_collector_env,
@@ -9,8 +9,12 @@ from minari.storage.hosting import (
     list_remote_datasets,
     upload_dataset,
 )
-from minari.storage.local import delete_dataset, list_local_datasets, load_dataset
-from minari.utils.data_collector import (
+from minari.storage.local import (
+    delete_dataset,
+    list_local_datasets,
+    load_dataset
+)
+from minari.data_collector import (
     DataCollectorV0,
     EpisodeMetadataCallback,
     StepDataCallback,
